requirements.txt
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0
